fixed broken unit tests

+++ src/structs/disease_progression.jl
@@ -189,34 +189,35 @@ struct DiseaseProgression
     end
 
     function DiseaseProgression(;
-        exposure::Int16 = Int16(-1),
-        infectiousness_onset::Int16 = Int16(-1),
-        symptom_onset::Int16 = Int16(-1),
-        severeness_onset::Int16 = Int16(-1),
-        hospital_admission::Int16 = Int16(-1),
-        icu_admission::Int16 = Int16(-1),
-        icu_discharge::Int16 = Int16(-1),
-        ventilation_admission::Int16 = Int16(-1),
-        ventilation_discharge::Int16 = Int16(-1),
-        hospital_discharge::Int16 = Int16(-1),
-        severeness_offset::Int16 = Int16(-1),
-        recovery::Int16 = Int16(-1),
-        death::Int16 = Int16(-1)
+        exposure::Integer = -1,
+        infectiousness_onset::Integer = -1,
+        symptom_onset::Integer = -1,
+        severeness_onset::Integer = -1,
+        hospital_admission::Integer = -1,
+        icu_admission::Integer = -1,
+        icu_discharge::Integer = -1,
+        ventilation_admission::Integer = -1,
+        ventilation_discharge::Integer = -1,
+        hospital_discharge::Integer = -1,
+        severeness_offset::Integer = -1,
+        recovery::Integer = -1,
+        death::Integer = -1
     )
+        
         return DiseaseProgression(
-            exposure,
-            infectiousness_onset,
-            symptom_onset,
-            severeness_onset,
-            hospital_admission,
-            icu_admission,
-            icu_discharge,
-            ventilation_admission,
-            ventilation_discharge,
-            hospital_discharge,
-            severeness_offset,
-            recovery,
-            death
+            Int16(exposure), 
+            Int16(infectiousness_onset), 
+            Int16(symptom_onset), 
+            Int16(severeness_onset), 
+            Int16(hospital_admission), 
+            Int16(icu_admission), 
+            Int16(icu_discharge), 
+            Int16(ventilation_admission), 
+            Int16(ventilation_discharge), 
+            Int16(hospital_discharge), 
+            Int16(severeness_offset), 
+            Int16(recovery), 
+            Int16(death)
         )
     end
 end
